% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpaddr_install.R
\name{jpaddr_install}
\alias{jpaddr_install}
\title{Install packages}
\source{
\href{https://github.com/geolonia/normalize-japanese-addresses}{geolonia/normalize-japanese-addresses}

\href{https://github.com/geolonia/japanese-addresses}{geolonia/japanese-addresses}
}
\usage{
jpaddr_install()
}
\description{
Install packages
}
