% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpaddr_npm_install.R
\name{jpaddr_npm_install}
\alias{jpaddr_npm_install}
\title{Install npm package}
\source{
\href{https://github.com/geolonia/normalize-japanese-addresses}{@geolonia/normalize-japanese-addresses}
}
\usage{
jpaddr_npm_install()
}
\description{
Install npm package
}
