% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_address.R
\name{normalize_address}
\alias{normalize_address}
\title{Normalize Japanese address}
\source{
\href{https://github.com/geolonia/normalize-japanese-addresses}{@geolonia/normalize-japanese-addresses}
}
\usage{
normalize_address(address, level = "town")
}
\arguments{
\item{address}{Address to normalize.}

\item{level}{Level of address normalization. One of \code{"town"}, \code{"city"}, or
\code{"pref"} (by default, \code{"town"}).}
}
\value{
A tibble with normalized address.
}
\description{
Normalize Japanese address
}
